move RowsIndexConfig
