+++ services/worker/tests/job_runners/config/test_parquet_metadata.py
@@ -318,6 +318,7 @@ def test_ParquetIndexWithMetadata_query(
         num_bytes=[num_bytes],
         httpfs=httpfs,
         hf_token=CI_USER_TOKEN,
+        max_arrow_data_in_memory=999999999,
     )
     with patch("libcommon.parquet_utils.HTTPFile", AuthenticatedHTTPFile):
         out = index.query(offset=0, length=2).to_pydict()
